add words extraction test
